wip: shot at abstracting request a bit cleaner
