+++ crates/unleash-api-client/src/http.rs
@@ -1,8 +1,10 @@
 // Copyright 2020, 2022 Cognite AS
 //! The HTTP layer.
 
+#[cfg(any(feature = "reqwest", feature = "reqwest-11", feature = "reqwest-13"))]
+pub mod reqwest;
 pub mod transport;
 
 #[cfg(any(feature = "reqwest", feature = "reqwest-11", feature = "reqwest-13"))]
-pub use transport::default_transport;
+pub use reqwest::default_transport;
 pub use transport::{BoxError, Http, Method, Request, Response, Transport, TransportRef};
