+++ crates/unleash-api-client/src/http/transport.rs
@@ -132,151 +132,6 @@ impl<T: Transport> Http<T> {
     }
 }
 
-#[cfg(feature = "reqwest")]
-pub struct ReqwestTransport {
-    client: reqwest::Client,
-}
-
-#[cfg(feature = "reqwest")]
-impl ReqwestTransport {
-    pub fn new(client: reqwest::Client) -> Self {
-        Self { client }
-    }
-}
-
-#[cfg(feature = "reqwest")]
-impl Default for ReqwestTransport {
-    fn default() -> Self {
-        Self::new(reqwest::Client::default())
-    }
-}
-
-#[cfg(feature = "reqwest")]
-#[async_trait]
-impl Transport for ReqwestTransport {
-    async fn execute(&self, request: Request) -> Result<Response, BoxError> {
-        let mut builder = match request.method {
-            Method::Get => self.client.get(&request.url),
-            Method::Post => self.client.post(&request.url),
-        };
-        for (name, value) in request.headers {
-            builder = builder.header(name, value);
-        }
-        if let Some(body) = request.body {
-            builder = builder
-                .header("content-type", "application/json")
-                .body(body);
-        }
-
-        let response = builder.send().await?;
-        let status = response.status().as_u16();
-        let body = response.bytes().await?.to_vec();
-
-        Ok(Response { status, body })
-    }
-}
-
-#[cfg(feature = "reqwest-11")]
-pub struct Reqwest11Transport {
-    client: reqwest_11::Client,
-}
-
-#[cfg(feature = "reqwest-11")]
-impl Reqwest11Transport {
-    pub fn new(client: reqwest_11::Client) -> Self {
-        Self { client }
-    }
-}
-
-#[cfg(feature = "reqwest-11")]
-impl Default for Reqwest11Transport {
-    fn default() -> Self {
-        Self::new(reqwest_11::Client::default())
-    }
-}
-
-#[cfg(feature = "reqwest-11")]
-#[async_trait]
-impl Transport for Reqwest11Transport {
-    async fn execute(&self, request: Request) -> Result<Response, BoxError> {
-        let mut builder = match request.method {
-            Method::Get => self.client.get(&request.url),
-            Method::Post => self.client.post(&request.url),
-        };
-        for (name, value) in request.headers {
-            builder = builder.header(name, value);
-        }
-        if let Some(body) = request.body {
-            builder = builder
-                .header("content-type", "application/json")
-                .body(body);
-        }
-
-        let response = builder.send().await?;
-        let status = response.status().as_u16();
-        let body = response.bytes().await?.to_vec();
-
-        Ok(Response { status, body })
-    }
-}
-
-#[cfg(feature = "reqwest-13")]
-pub struct Reqwest13Transport {
-    client: reqwest_13::Client,
-}
-
-#[cfg(feature = "reqwest-13")]
-impl Reqwest13Transport {
-    pub fn new(client: reqwest_13::Client) -> Self {
-        Self { client }
-    }
-}
-
-#[cfg(feature = "reqwest-13")]
-impl Default for Reqwest13Transport {
-    fn default() -> Self {
-        Self::new(reqwest_13::Client::default())
-    }
-}
-
-#[cfg(feature = "reqwest-13")]
-#[async_trait]
-impl Transport for Reqwest13Transport {
-    async fn execute(&self, request: Request) -> Result<Response, BoxError> {
-        let mut builder = match request.method {
-            Method::Get => self.client.get(&request.url),
-            Method::Post => self.client.post(&request.url),
-        };
-        for (name, value) in request.headers {
-            builder = builder.header(name, value);
-        }
-        if let Some(body) = request.body {
-            builder = builder
-                .header("content-type", "application/json")
-                .body(body);
-        }
-
-        let response = builder.send().await?;
-        let status = response.status().as_u16();
-        let body = response.bytes().await?.to_vec();
-
-        Ok(Response { status, body })
-    }
-}
-
-#[cfg(any(feature = "reqwest", feature = "reqwest-11", feature = "reqwest-13"))]
-pub fn default_transport() -> TransportRef {
-    cfg_if::cfg_if! {
-        if #[cfg(feature = "reqwest")] {
-            Arc::new(ReqwestTransport::default())
-        } else if #[cfg(feature = "reqwest-11")] {
-            Arc::new(Reqwest11Transport::default())
-        } else if #[cfg(feature = "reqwest-13")] {
-            Arc::new(Reqwest13Transport::default())
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
